tg_bots/requirements.txt
aiogram==2.23.1
telethon==1.24.0
python-dotenv==1.0.0
requests==2.28.1